Data_Adjuster/requirements.txt
pandas>=1.3.0
numpy>=1.21.0
pyyaml>=5.4
